+++ packages/headless/src/features/facets/automatic-facet-set/interfaces/request.ts
@@ -1,0 +1,7 @@
+import {FacetValue} from '../../../../product-listing.index';
+
+export type AutomaticFacetRequest = {
+  field: string;
+  label: string;
+  currentValues: FacetValue[];
+};