+++ packages/headless/src/features/facets/facet-set/interfaces/response.ts
@@ -5,8 +5,3 @@ export interface FacetValue extends BaseFacetValue {
 }
 
 export type FacetResponse = BaseFacetResponse<FacetValue>;
-
-export type AutomaticFacetResponse = Omit<
-  BaseFacetResponse<FacetValue>,
-  'facetId'
->;