+++ packages/quantic/force-app/examples/main/lwc/exampleQuanticRecommendationList/exampleQuanticRecommendationList.js
@@ -32,6 +32,8 @@ export default class ExampleQuanticRecommendationtList extends LightningElement 
       attribute: 'label',
       label: 'Label',
       description: 'The label of the component.',
+      defaultValue:
+      'Top documents for you',
     },
     {
       attribute: 'fieldsToInclude',
@@ -44,8 +46,14 @@ export default class ExampleQuanticRecommendationtList extends LightningElement 
     {
       attribute: 'headingLevel',
       label: 'Heading Level',
-      description: 'The Heading level to use for the heading label',
+      description: 'The Heading level to use for the heading label. Will display the label of the recommendations within a <hx> tag.',
       defaultValue: 1,
+    },
+    {
+      attribute: 'variant',
+      label: 'Variant',
+      description: "The variant of the component. Accepted variants are `grid` and `carousel`.",
+      defaultValue: 'grid',
     },
   ];
   expectedEvents = ['registerrecommendationtemplates'];