+++ packages/quantic/force-app/main/default/lwc/quanticInsightInterface/quanticInsightInterface.js
@@ -39,9 +39,20 @@ export default class QuanticInsightInterface extends LightningElement {
   engineOptions;
   /** @type {boolean} */
   initialized;
+  /** @type {boolean} */
+  hasRendered = false;
+  /** @type {boolean} */
+  ariaLiveEventsBound = false;
 
   disconnectedCallback() {
     destroyEngine(this.engineId);
+    if (this.ariaLiveEventsBound) {
+      this.removeEventListener('arialivemessage', this.handleAriaLiveMessage);
+      this.removeEventListener(
+        'registerregion',
+        this.handleRegisterAriaLiveRegion
+      );
+    }
   }
 
   connectedCallback() {
@@ -70,6 +81,13 @@ export default class QuanticInsightInterface extends LightningElement {
     });
   }
 
+  renderedCallback() {
+    if (!this.hasRendered && this.querySelector('c-quantic-aria-live')) {
+      this.bindAriaLiveEvents();
+    }
+    this.hasRendered = true;
+  }
+
   initialize = () => {
     if (this.initialized) {
       return;
@@ -92,5 +110,40 @@ export default class QuanticInsightInterface extends LightningElement {
    */
   get input() {
     return this.template.querySelector('input');
+  }
+
+  bindAriaLiveEvents() {
+    this.template.addEventListener(
+      'arialivemessage',
+      this.handleAriaLiveMessage.bind(this)
+    );
+    this.template.addEventListener(
+      'registerregion',
+      this.handleRegisterAriaLiveRegion.bind(this)
+    );
+    this.ariaLiveEventsBound = true;
+  }
+
+  handleAriaLiveMessage(event) {
+    /** @type {import('quanticAriaLive/quanticAriaLive').IQuanticAriaLive} */
+    const ariaLiveRegion = this.querySelector('c-quantic-aria-live');
+    if (ariaLiveRegion) {
+      ariaLiveRegion.updateMessage(
+        event.detail.regionName,
+        event.detail.message,
+        event.detail.assertive
+      );
+    }
+  }
+
+  handleRegisterAriaLiveRegion(event) {
+    /** @type {import('quanticAriaLive/quanticAriaLive').IQuanticAriaLive} */
+    const ariaLiveRegion = this.querySelector('c-quantic-aria-live');
+    if (ariaLiveRegion) {
+      ariaLiveRegion.registerRegion(
+        event.detail.regionName,
+        event.detail.assertive
+      );
+    }
   }
 }